refactor: types in types.tsx
